Meme Generator (1)/Meme Generator/requirements.txt
streamlit 
transformers 
torch 
diffusers 
pillow 
requests
mediapipe
torchvision 
opencv-python